requirements.txt
loguru==0.7.0
pandas==2.0.3
PyMySQL==1.1.0
python-dotenv==1.0.1
pywinauto==0.6.8
Requests==2.31.0
selenium==4.17.2
